fix(connect sdk): Split up html and js and split off snippets
I split up the html and js file so you can see the js file growing step per step and this way it's easier to copy paste without incorrect indentation.

The snippets code, just like the boilerplate and demos, is really important and therefore I think it deserves it's own section. It also has very little to do with connecting the SDK and seems misplaced here.